Revert "Add test coverage for a svcat plan update without a plan"
This reverts commit 0ec6358b

Fix in 2a0f8d9 was tested manually with svcat configured with a controller with 3 replicas.

The test isn't reproducing the problem fixed in 2a0f8d9 because it simulating an IN-PROGRESS conflict while 2a0f8d9 fixed an SUCCESS conflict.

Covering the SUCCESS conflict in UpdateInstanceWithBackingServiceAcceptanceTest is costly to implement since likelyhood of managing to reproduce synchronously such conflict is low. E2E test for this should rather use svcat with multiple replicas.

Also, lowering the priority of covering svcat concurrent calls due to low probability of occurence of them due to https://github.com/kubernetes-sigs/service-catalog/issues/2888